# Conflicts:
#	curriculum-server/pom.xml
#	curriculum-server/src/main/java/com/curriculum/service/OrderMainService.java
#	curriculum-server/src/main/java/com/curriculum/service/ShoppingCartService.java
#	curriculum-server/src/main/java/com/curriculum/service/impl/OrderMainServiceImpl.java
#	curriculum-server/src/main/java/com/curriculum/service/impl/OrdersServiceImpl.java
#	curriculum-server/src/main/java/com/curriculum/service/impl/ShoppingCartServiceImpl.java

+++ curriculum-base/src/main/java/com/curriculum/model/po/PayRecord.java
@@ -53,7 +53,7 @@ public class PayRecord implements Serializable {
     /**
      * 订单总价单位元
      */
-    private Float totalPrice;
+    private Double totalPrice;
 
     /**
      * 币种CNY
